requirements.txt
couchdb >= 0.8
py-trello
oauth2
flowcell_parser